moved OutStream out in to seprate file
